refactor/EV-870/refactoring files and folders for test
