+++ test/tests/shared/shared_widgets_test.dart
@@ -1,9 +1,10 @@
 import 'package:everest_flutter_crypto_tests/shared/widgets/bottom_nav_bar.dart';
 import 'package:everest_flutter_crypto_tests/shared/widgets/default_appbar.dart';
+import 'package:faker/faker.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import 'helpers/setup_widget_tester.dart';
+import '../../helpers/setup_widget_tester.dart';
 
 class Counter {
   int value = 0;
@@ -27,8 +28,8 @@ main() {
         "WHEN DefaultApp receives information, THEN ensure that the information is being used",
         (WidgetTester tester) async {
       final counter = Counter();
-      const IconData specificIcon = Icons.abc;
-      const String specificTitle = '';
+      IconData specificIcon = Icons.abc;
+      String specificTitle = faker.guid.guid();
 
       await loadPage(
         tester,
@@ -65,6 +66,4 @@ main() {
       expect(find.byKey(const Key('cryptoCurrencyIcon')), findsOneWidget);
     });
   });
-
-
 }
